исправление ошибок
